Deleted code that creates swap file. Replaced it with simple swap size check
While the code from the previous commit probably works, it appears
that hibernating to a swap file instead of a swap partition is harder
than I originally thought. Not sure why I didn't test this before writing
the code (or did I?). Uswsusp might be able to do it, but that requires
additional code, and testing. For now, I removed the swap file creation
code, replacing it with a simple check to see if we have enough swap
space to hibernate.